Linux warning fixes (#121)
* Update GStreamerVideo.cpp

linux warnings

* more GStreamer warnings

* Update Configuration.cpp

* Update CollectionInfo.cpp

* Update PageBuilder.cpp

* Update PageBuilder.cpp

try 2

* Update AttractMode.cpp

* Update Tween.cpp

* Update Page.cpp

* Update ReloadableMedia.cpp

* Update CMakeLists.txt

suppress reorder warnings

---------

+++ RetroFE/Source/Graphics/Component/ReloadableMedia.cpp
@@ -360,7 +360,7 @@ void ReloadableMedia::reloadTexture()
         else if (typeLC == "position")
         {
             if (!selectedItem->collectionInfo->items.empty()) {
-                int position = page.getSelectedIndex() + 1;
+                size_t position = page.getSelectedIndex() + 1;
                 if (position == 1) {
                     basename = '1';
                 }
